Move operators/utils to internals
