+++ packages/confluence/src/getPageAttachments.spec.ts
@@ -1,5 +1,5 @@
 import url from 'url'
-import { promisify } from '@etlx/operators/utils'
+import { promisify } from '@etlx/operators/@internal/utils'
 import { mockFetch, jsonResponse } from '@etlx/operators/@internal/testing/fetch'
 import { confluence, page, dataPage, respondWith } from './@internal/testing'
 import { getPageAttachments } from './getPageAttachments'
