+++ packages/confluence/src/getSpaceContent.ts
@@ -1,7 +1,7 @@
 import { of, Observable } from 'rxjs'
 import { mergeMap, map } from 'rxjs/operators'
 import { expandWhile } from '@etlx/operators/core'
-import { isNullOrUndefined } from '@etlx/operators/utils'
+import { isNullOrUndefined } from '@etlx/operators/@internal/utils'
 import { log, Logger } from '@etlx/operators/@internal/log'
 import { fromJsonRequest, FromRequestOptions, authBasic } from '@etlx/operators/http'
