+++ packages/operators/src/@internal/utils/array.ts
