+++ packages/operators/src/core/throwWhen.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs'
 import { tap } from 'rxjs/operators'
-import { not, throwError } from '../utils'
+import { not, throwError } from '../@internal/utils'
 
 export const throwWhen = <T>(predicate: (x: T) => boolean, error: (x: T) => string | Error) =>
     ($: Observable<T>) => $.pipe(
